[CHORE] data class 합치기
