--- app/src/main/java/sopt/haeti/baeminaos/data/local/StoreMenuDetail.kt
@@ -1,8 +1,0 @@
-package sopt.haeti.baeminaos.data.local
-
-data class StoreMenuDetail(
-    val menuDetailName: String,
-    val menuDescription: String,
-    val menuPrice: Int,
-    val menuImage: String,
-)
